FEAT: add simple fetch service
The FetchBinaryService is added. This service is able to fetch the node
binary according to the given version and flavor. Extra modals have been
added to parse the appsettings and map the response of the api call.

A retry has been added as well, this makes it possible to retry multiple
times before failing the request.